+++ App.js
@@ -3,7 +3,7 @@ import { NavigationContainer } from "@react-navigation/native";
 import NavStacks from "./src/navigation/NavStacks";
 import { TailwindProvider } from "tailwindcss-react-native";
 import { Provider } from "react-redux";
-import store from "./src/redux/store";
+import store from "./src/redux/store/store";
 
 export default function App() {
   return (
