+++ src/components/teams/Teams.js
@@ -9,7 +9,7 @@ import {
 import TeamsCard from "./TeamsCard";
 
 import { useDispatch, useSelector } from "react-redux";
-import { fetchTeams, selectTeamItems } from "../../redux/teamSlice";
+import { fetchTeams, selectTeamItems } from "../../redux/slices/teamSlice";
 
 const Teams = () => {
   const teams = useSelector(selectTeamItems);
